requirements.txt
python-dotenv==1.2.1
slack_bolt==1.27.0
slack_sdk==3.39.0
